strip microseconds from last reboot (#2)
* strip microseconds from last reboot

* Update __init__.py

* Update __init__.py

Keep track of previous Last Reboot value, and only change the state if the new calculation is different by more than 1.0 seconds

* Update __init__.py

* Set Last Reboot state change threshold to 5 sec

I tested this on a Pi for a couple of days, to check slower system behavior. I think it fixes that last issue, where once in a while it was still changing the state value without a reboot. The state doesn't update at all with this 5 second threshold, unless the actual Obi is rebooted. I don't think it's realistic that Obi devices would complete a reboot cycle within 5 seconds, so this limit should work fine. 

One caveat, restarting HA can sometimes result in a one-time Last Reboot state value change -- as observed here https://github.com/dshokouhi/pyobihai/pull/2#issuecomment-632947629 -- even when the Obi hasn't been rebooted. This is a bootstrapping problem; we don't pass the current state value into the PyObihai object when it is created. I think this behavior is OK.